get rid of animation folder
