--- src/shared/components/Animations/list-item.less
@@ -1,69 +1,0 @@
-@import '../../lib.less';
-
-// mixin
-
-.listItemLike() {
-  .unstyleList();
-  background-color: @component-background;
-  border-radius: 2px;
-  cursor: pointer;
-  margin: 0.5em 0;
-  padding: 0.5em;
-  position: relative;
-}
-
-.list-item {
-  .listItemLike();
-  display: flex;
-  flex-direction: column;
-  &:focus,
-  &:focus-within {
-    background-color: fade(@primary-color, 5%);
-  }
-  &:focus {
-    .focus();
-  }
-  &:hover > .content > .actions {
-    display: block;
-  }
-  > .image-preview > .wrapper > .skeleton {
-    width: 100%;
-    height: 200px;
-  }
-  > .content {
-    display: flex;
-    justify-content: space-between;
-    height: 24px;
-    flex-grow: 1;
-  }
-  > .content > .label {
-    .ellipsis();
-    color: @text-color;
-    font-weight: 500;
-    flex-grow: 1;
-    margin-right: @default-pad;
-  }
-  > .content > .label > .predicate {
-    font-size: 0.8em;
-    opacity: 0.5;
-    margin-right: 0.5em;
-  }
-  > .content > .details {
-    display: flex;
-  }
-  > .content > .actions {
-    margin-left: 0.5em;
-    display: none;
-  }
-  > .description {
-    margin: 0;
-    color: @text-color-secondary;
-  }
-  &.-compact > .description {
-    z-index: 1;
-    .ellipsis();
-  }
-  &.-small {
-    flex-direction: row;
-  }
-}
